+++ internal/app/repo/storage_repo.go
@@ -22,22 +22,10 @@ type repo struct {
 }
 
 // NewRepo returns Repo interface
-func NewRepo(db *sqlx.DB, batchSize uint) Repo {
+func NewRepo(db *sqlx.DB, batchSize uint) *repo {
 	return &repo{db: db, batchSize: batchSize}
 }
 
 func (r *repo) Describe(ctx context.Context, templateID uint64) (*model.Package, error) {
 	return nil, nil
-}
-
-type EventRepo interface {
-	// Lock заблокировать в БД n записей
-	Lock(n uint64) ([]model.PackageEvent, error) // TODO в имплементации реализовать загрузку зависших залоченных событий, с отсечкой по времени
-	// Unlock разблокировать в БД n записей
-	Unlock(eventID []uint64) error
-
-	// Add добавить в БД n записей
-	Add(event []model.PackageEvent) error
-	// Remove удалить из БД n записей
-	Remove(eventIDs []uint64) error
 }
