+++ internal/model/errors.go
@@ -1,0 +1,7 @@
+package model
+
+import "errors"
+
+var (
+	ErrNotImplemented = errors.New("not implemented")
+)
